// src/app/contracts/enums/notification.enum.ts
export enum NotificationAction {
  FORGOT_PASSWORD = 'forgot_password',
  FRANCHISE_ADMIN_CREATED = 'franchise_admin_created',
  SERVICE_TYPE_CATEGORY_REQUESTED = 'service_type_category_requested',
  SERVICE_TYPE_REQUESTED = 'service_type_requested',
  SERVICE_TYPE_CATEGORY_CREATED = 'service_type_category_created',
  SERVICE_TYPE_CATEGORY_DELETED = 'service_type_category_deleted',
  SERVICE_TYPE_REQUEST_CREATED = 'service_type_request_created',
  STANDARD_ADMIN_CREATED = 'standard_admin_created',
  STANDARD_ADMIN_UPDATED = 'standard_admin_updated',
  VENDOR_CREATED = 'vendor_created',
  VENDOR_STATUS_CHANGED = 'vendor_status_changed',
  ESTIMATE_CREATED = 'estimate_created',
  ESTIMATE_REJECTION_BY_OWNER_FOR_VENDOR = 'estimate_rejection_by_owner_for_vendor',
  ESTIMATE_REJECTION_BY_OWNER_FOR_FRANCHISE_ADMIN = 'estimate_rejection_by_owner_for_fa',
  ESTIMATE_VENDOR_ASSIGNMENT_SINGLE = 'estimate_vendor_assignment_single',
  ESTIMATE_VENDOR_ASSIGNMENT_MULTI = 'estimate_vendor_assignment_multi',
  ESTIMATE_VENDOR_QUOTATION_UPDATE_BY_FRANCHISE_ADMIN = 'estimate_vendor_quotation_update_by_fa',
  ESTIMATE_VENDOR_QUOTATION_ADDED = 'estimate_vendor_quotation_added',
  INVOICE_REJECTED = 'invoice_rejected',
  INVOICE_CREATED = 'invoice_created',
  INVOICE_UPDATED = 'invoice_updated',
  INVOICE_SENT_TO_OWNER = 'invoice_sent_to_owner',
  INVOICE_PAYMENT_RECORDED_BY_FRANCHISE_ADMIN = 'invoice_payment_recorded_by_franchise_admin',
  INVOICE_PAYMENT_MADE_BY_OWNER = 'invoice_payment_made_by_owner',
  INVOICE_PAYMENT_FAILED_BY_OWNER = 'invoice_payment_failed_by_owner',
  INVOICE_PAYMENT_INITIATED_BY_OWNER = 'invoice_payment_initiated_by_owner',
  OWNER_CREATED = 'owner_created',
  SERVICE_REQUEST_CREATED = 'service_request_created',
  SERVICE_REQUEST_RELEASE_VENDOR = 'service_request_release_created',
  SERVICE_REQUEST_STATUS_UPDATE = 'service_request_status_update',
  SERVICE_REQUEST_CLAIMED = 'service_request_claimed',
  SERVICE_REQUEST_DISCREPANCY_CREATED = 'service_request_discrepancy_claimed',
  SERVICE_REQUEST_APPROVAL = 'service_request_approval',
  OWNER_ARCHIVED = 'owner_archived',
  OWNER_UNARCHIVE = 'owner_unarchive',
  SERVICE_REQUEST_UPDATE = 'service_request_update',
  VENDOR_UNASSIGNED = 'vendor_unassigned',
  VENDOR_ASSIGNED = 'vendor_assigned',
  BULK_VENDOR_ASSIGNED = 'bulk_vendor_assigned',
  SERVICE_REQUEST_CANCEL = 'service_request_cancel',
  MEMBERSHIP_PAYMENT_FAILED = 'membership_payment_failed',
  MEMBERSHIP_PAYMENT_SUCCESS = 'membership_payment_success',
  PAYMENT_METHOD_ATTACHED_SUCCESS = 'payment_method_attached_success',
  PAYMENT_METHOD_ATTACHED_FAILED = 'payment_method_attached_failed',
}